Add multiple widgets with the same and different size
